requirements.txt
streamlit
pandas
python-docx
python-dotenv
openai
